[sqlite][ios] throw custom error when can't open shared container [sqlite][android] adjust native code signatures [core][ios] add custom handling for permission check of path in a shared container [file-system][ios] add getSharedContainerUriAsync [sqlite][ios] add tests for app group use [file-system][ios] add tests for getSharedContainerUriAsync use

+++ packages/expo-sqlite/build/ExpoSQLiteNext.d.ts
@@ -2,6 +2,8 @@ import type { SQLiteOpenOptions, IOSOptions } from './NativeDatabase';
 import type { DatabaseChangeEvent } from './SQLiteDatabase';
 declare const _default: {
     NativeDatabase(databaseName: string, iosOptions?: IOSOptions, options?: SQLiteOpenOptions, serializedData?: Uint8Array): void;
+    ensureHasAccessAsync(databaseName: string, iosOptions?: IOSOptions): Promise<void>;
+    ensureHasAccessSync(databaseName: string, iosOptions?: IOSOptions): void;
     NativeStatement(): void;
     deleteDatabaseAsync(databaseName: string, iosOptions: IOSOptions): Promise<void>;
     deleteDatabaseSync(databaseName: string, iosOptions: IOSOptions): void;
